pyproject.toml
[tool.poetry]
name = "family-chat"
version = "0.1.0"
description = "Fun and helpful chatbot"
authors = ["DrDavidL <92898146+DrDavidL@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10,<3.12"
openai = "^1.25.1"
streamlit = "^1.34.0"
embedchain = "^0.1.103"
pysqlite3 = "^0.5.2"
langchain-text-splitters = "^0.0.1"
requests = "^2.31.0"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
